Fixed current Notification Type not being set
Currently it does not throw an error when SMS is selected with phone number.  That might be a backend issue.  There is also no error thrown when Push Notification is selected when there is no firebase id.  This might be a front end issue since it does not update the user if there is no firebase id.